Fix overflow bug

+++ contracts/staking-positions/PangolinStakingPositionsFunding.sol
@@ -138,7 +138,7 @@ abstract contract PangolinStakingPositionsFunding is AccessControlEnumerable, Ge
             }
         } else {
             unchecked {
-                uint256 leftover = (periodFinish - lastUpdate) * _rewardRate;
+                uint256 leftover = (uint256(periodFinish) - lastUpdate) * _rewardRate;
                 assembly {
                     newRewardRate := div(add(amount, leftover), tmpPeriodDuration)
                 }
